+++ shared/main/java/net/ccbluex/liquidbounce/utils/Rotation.kt
@@ -155,8 +155,8 @@ data class Rotation(var yaw: Float, var pitch: Float) : MinecraftInstance()
 			val yawSin = WMathHelper.sin(WMathHelper.toRadians(yaw))
 			val yawCos = WMathHelper.cos(WMathHelper.toRadians(yaw))
 
-			thePlayer.motionX += calcStrafe * yawCos - calcForward * yawSin.toDouble()
-			thePlayer.motionZ += calcForward * yawCos + calcStrafe * yawSin.toDouble()
+			thePlayer.motionX += calcStrafe * yawCos - calcForward * yawSin
+			thePlayer.motionZ += calcForward * yawCos + calcStrafe * yawSin
 		}
 	}
 }
